Introduces flattening of types without a known type editor.
Types which do not have a known type editor are browsed and flattened into the highest level as if their properties belong in the class that is currently handled by the PropertyGrid.
Furthermore some flatting options are available to decide in which category to place the flattened properties and how to name that category.
While the flattening is off by default to keep the default behavior, the demo has been extented to do apply flattening.