Introduces a dependency property ShowSearchBar in order to hide the search bar. The search bar XAML mark up has been updated to adjust the visibility accordingly.
Furthermore, multibinding is used to collapse the dockpanel visbility if the search bar and the sort buttons are both hidden.

To support that the BooleanArr2VisibilityConverter has been extented to optionally return a visible state if any of the bindings are visible (instead of all).
For that the ConverterParameter can be set to 'UseAny'.